Export launcher script to TypeScript
